+++ Properties/AssemblyInfo.cs
@@ -5,8 +5,8 @@ using System.Runtime.Versioning;
 
 [assembly: AssemblyCompany("NiumaClaw.Team")]
 [assembly: AssemblyConfiguration("Release")]
-[assembly: AssemblyFileVersion("1.0.22.0")]
-[assembly: AssemblyInformationalVersion("1.0.22+desktop-cli-path.20260702")]
+[assembly: AssemblyFileVersion("1.0.23.0")]
+[assembly: AssemblyInformationalVersion("1.0.23+runner-worklog-history.20260702")]
 [assembly: AssemblyProduct("NiumaClaw.Team")]
 [assembly: AssemblyTitle("NiumaClaw.Team")]
-[assembly: AssemblyVersion("1.0.22.0")]
+[assembly: AssemblyVersion("1.0.23.0")]